feat: Add event model
This commit adds a new file, `event.model.ts`, which defines the event model schema using Mongoose. The event model includes fields for `title`, `description`, `date`, `location`, `capacity`, `organizers`, and `speakers`. The `title` field is required. The `description` field has a minimum length of 40 characters and is also required. The `date` field has a default value of the current date plus 24 hours and is required. The `location` field is required. The `capacity` field has a default value of 10 and is required to be at least 10. The `organizers` and `speakers` fields are arrays of `User` objects referenced by their `ObjectId`.