feat: Add ticket model
This commit adds a new file, `ticket.model.ts`, which defines the ticket model schema using Mongoose. The ticket model includes fields for `event`, `type`, `price`, `quantity`, and `buyer`. The `event` field is a reference to the `Event` model. The `type` field is an array that is yet to be decided. The `price` field is required. The `quantity` field is required. The `buyer` field is a reference to the `User` model and is required.

Co-authored-by: Rohit