feat: Add Toaster component to RootLayout
The code changes include adding the Toaster component to the RootLayout in order to display toast notifications. This will provide a user-friendly way to show important messages or alerts to the user. The Toaster component will be rendered as a child of the ThemeProvider component, ensuring that it is available throughout the application. This commit follows the established convention of starting with a verb in the present tense and providing a concise description of the changes made.